refactor: rename module matching their feature
